Refactor based on comments by andrew hill.
[INT-9647]
