+++ src/domain/Internals/Collapsible/index.ts
@@ -1,0 +1,1 @@
+export { Collapsible, IToggleType } from './Collapsible'
