+++ src/domain/Internals/Collapsible/style.ts
