--- src/domain/Menus/Menu/Collapsible/index.ts
@@ -1,1 +1,0 @@
-export { Collapsible } from './Collapsible'
