+++ src/domain/Menus/Menu/MenuItem/MenuItemContent/index.ts
@@ -1,0 +1,1 @@
+export { MenuItemContent } from './MenuItemContent'
